requirements.txt
#Para instalar todos los paquees pip: -r requirements.txt
django
django-admin
dj_rest_auth
django-allauth
psycopg2-binary
pillow
django-rest-framework
django-cors-headers
django-countries[pyuca]
django-simple-history
drf-yasg #documentación API Swagger
django_seed


